realtime Chat Box completed
worked on UI and backend of chat feature. Stored chats in database and retrieved them using php. Such that no chat is repeated and no chat sent by an user should appear to be sent by another user or in short checked authentication and ids of messages and ensure that they are delivered properly